pyproject.toml
[project]
name = "youtube-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.193.0",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
    "python-mpv>=1.0.8",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.3.17",
]
